+++ cloud-siem-20220616/ChangeLog.md
@@ -1,3 +1,9 @@
+2026-09-14 Version: 6.3.0
+- Support API DescribeUserSiemOrderStatus.
+- Update API DescribeAlertsWithEvent: add response parameters Body.Data.ResponseData.$.AlertStatus.
+- Update API PostEventDisposeAndWhiteruleList: add request parameters SyncAlertStatus.
+
+
 2026-08-31 Version: 6.2.4
 - Update API EnableAccessForCloudSiem: add request parameters ClientToken.
 - Update API PostEventDisposeAndWhiteruleList: add request parameters ClientToken.